backend/requirement.txt
langgraph
langchain
langchain-core
langchain_community
chromadb
tavily-python
 requests
 sqlalchemy
 pydantic
 fastapi
 uvicorn
 python-dotenv